+++ likebutton.js
@@ -1,4 +1,4 @@
-import {renderComments} from "./main.js"
+import {renderComments} from "./renderComments.js"
 
 function delay(interval = 300) {
     return new Promise((resolve) => {
@@ -19,7 +19,7 @@ export const initLikeButtonListener = (comments) => {
           comments[index].like_active
             ? comments[index].like_count++
             : comments[index].like_count--;
-          renderComments();
+          renderComments(comments);
         })
         .then(() => {
           el.classList.remove("-loading-like");
